move tab functions to tab context
